(Run + Converter) add toggle for main wrap
When using the new flag '-e', PYST will not wrap the generated user code in a main function. This allows the user to easily export the generated file as a module (provide). When running the generated file in standalone mode however, some extra (grabish) output might be generated (unused top-level function return values).